pyproject.toml
[project]
name = "nonebot-plugin-nonechat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9, <4.0"
dependencies = [
    "langchain-core>=0.3.6",
    "langchain-openai>=0.2.1",
    "langgraph-checkpoint-sqlite>=1.0.4",
    "langgraph>=0.2.28",
    "nonebot-adapter-onebot>=2.1.0",
    "nonebot-plugin-alconna>=0.52.3",
    "nonebot-plugin-htmlrender>=0.3.5",
    "nonebot-plugin-localstore>=0.7.1",
    "nonebot2[fastapi]>=2.3.3",
    "pydantic>=1.0,<3.0.0",
]

[tool.ruff]
ignore = ["E402", "F401"]

[tool.ruff.format]
quote-style = 'single'


[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]
plugins = []
plugin_dirs = ["nonebot_plugin_nonechat"]
builtin_plugins = []

[tool.uv]
dev-dependencies = [
    "mypy>=1.11.2",
    "nonebot-adapter-onebot>=2.4.5",
]
